refactor: Split date_arithmetic_tests.rs into focused modules
Refactor the 1,086-line date_arithmetic_tests.rs file into 5 focused modules
organized by operation type:

- date_operations.rs (193 lines): DATEDIFF, DATE_ADD, DATE_SUB with DATE types
- time_operations.rs (77 lines): Time-specific calculations (HOUR, MINUTE, SECOND)
- timestamp_operations.rs (52 lines): Timestamp with time component operations
- interval_operations.rs (138 lines): EXTRACT and AGE functions
- edge_cases.rs (636 lines): Leap years, month/year boundaries, null handling, errors

All 61 existing tests preserved and passing. Improves test organization and
maintainability by grouping related tests into logical categories.
